pyproject.toml
[project]
name = "kb-search-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.45.0",
    "uvicorn>=0.34.2",
]
